Creates two servlets to handle the two different proccesses for editing customers info

+++ src/com/webappcafe/service/CustomerService.java
@@ -26,7 +26,7 @@ public class CustomerService {
         
     }
     
-    public void editCustomerPassword(int id, String password) {
+    public void editCustomerPassword(long id, String password) {
         
         customerDAO = new CustomerDAOImpl();
